# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * F1 Dex admin API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type DriverRarity = typeof DriverRarity[keyof typeof DriverRarity];


export const DriverRarity = {
  common: 'common',
  rare: 'rare',
  epic: 'epic',
  legendary: 'legendary',
} as const;

export interface Driver {
  id: number;
  name: string;
  shortName: string;
  number: number;
  team: string;
  nationality: string;
  rarity: DriverRarity;
  rating: number;
  active: boolean;
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  spawnImageUrl: string | null;
  /** @minimum 0 */
  spawnWeight: number;
  createdAt: string;
}

export type DriverInputRarity = typeof DriverInputRarity[keyof typeof DriverInputRarity];


export const DriverInputRarity = {
  common: 'common',
  rare: 'rare',
  epic: 'epic',
  legendary: 'legendary',
} as const;

export interface DriverInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 2 */
  shortName: string;
  /** @minimum 0 */
  number: number;
  /** @minLength 2 */
  team: string;
  /** @minLength 2 */
  nationality: string;
  rarity: DriverInputRarity;
  /**
     * @minimum 1
     * @maximum 100
     */
  rating: number;
  active: boolean;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  spawnImageUrl?: string | null;
  /** @minimum 0 */
  spawnWeight?: number;
}

export type DriverUpdateRarity = typeof DriverUpdateRarity[keyof typeof DriverUpdateRarity];


export const DriverUpdateRarity = {
  common: 'common',
  rare: 'rare',
  epic: 'epic',
  legendary: 'legendary',
} as const;

export interface DriverUpdate {
  /** @minLength 2 */
  name?: string;
  /** @minLength 2 */
  shortName?: string;
  /** @minimum 0 */
  number?: number;
  /** @minLength 2 */
  team?: string;
  /** @minLength 2 */
  nationality?: string;
  rarity?: DriverUpdateRarity;
  /**
     * @minimum 1
     * @maximum 100
     */
  rating?: number;
  active?: boolean;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  spawnImageUrl?: string | null;
  /** @minimum 0 */
  spawnWeight?: number;
}

export type PlayerStatus = typeof PlayerStatus[keyof typeof PlayerStatus];


export const PlayerStatus = {
  active: 'active',
  flagged: 'flagged',
  inactive: 'inactive',
} as const;

export interface Player {
  id: number;
  username: string;
  discordTag: string;
  collectionCount: number;
  balance: number;
  status: PlayerStatus;
  lastActiveAt: string;
  favoriteDriver: string;
}

export type ActivityType = typeof ActivityType[keyof typeof ActivityType];


export const ActivityType = {
  drop: 'drop',
  trade: 'trade',
  player: 'player',
  race: 'race',
  system: 'system',
} as const;

export interface Activity {
  id: number;
  type: ActivityType;
  title: string;
  detail: string;
  timestamp: string;
  accent: string;
}

export interface DashboardSummary {
  totalPlayers: number;
  activeGuilds: number;
  cardsCollected: number;
  tradesToday: number;
  weeklyClaims: number;
  weeklyClaimsChange: number;
  collectionCompletion: number;
  topDriver: string;
  recentActivity: Activity[];
}

export interface BotStatus {
  connected: boolean;
  latencyMs: number;
  guildCount: number;
  version: string;
  branch: string;
  lastHeartbeat: string;
}

export type BotActionInputAction = typeof BotActionInputAction[keyof typeof BotActionInputAction];


export const BotActionInputAction = {
  'sync-drivers': 'sync-drivers',
  'publish-drop': 'publish-drop',
  'refresh-cache': 'refresh-cache',
  'pause-drops': 'pause-drops',
} as const;

export interface BotActionInput {
  action: BotActionInputAction;
}

export interface BotActionResult {
  success: boolean;
  action: string;
  message: string;
  timestamp: string;
}

export type GetDriversParams = {
search?: string;
};

export type GetPlayersParams = {
search?: string;
status?: GetPlayersStatus;
};

export type GetPlayersStatus = typeof GetPlayersStatus[keyof typeof GetPlayersStatus];


export const GetPlayersStatus = {
  active: 'active',
  flagged: 'flagged',
  inactive: 'inactive',
} as const;