# lib/api-zod/src/generated/types/driver.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * F1 Dex admin API
 * OpenAPI spec version: 0.1.0
 */
import type { DriverRarity } from './driverRarity';

export interface Driver {
  id: number;
  name: string;
  shortName: string;
  number: number;
  team: string;
  nationality: string;
  rarity: DriverRarity;
  rating: number;
  active: boolean;
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  spawnImageUrl: string | null;
  /** @minimum 0 */
  spawnWeight: number;
  createdAt: string;
}