# lib/api-zod/src/generated/types/driverInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * F1 Dex admin API
 * OpenAPI spec version: 0.1.0
 */
import type { DriverInputRarity } from './driverInputRarity';

export interface DriverInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 2 */
  shortName: string;
  /** @minimum 0 */
  number: number;
  /** @minLength 2 */
  team: string;
  /** @minLength 2 */
  nationality: string;
  rarity: DriverInputRarity;
  /**
     * @minimum 1
     * @maximum 100
     */
  rating: number;
  active: boolean;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  spawnImageUrl?: string | null;
  /** @minimum 0 */
  spawnWeight?: number;
}