# lib/api-zod/src/generated/types/driverUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * F1 Dex admin API
 * OpenAPI spec version: 0.1.0
 */
import type { DriverUpdateRarity } from './driverUpdateRarity';

export interface DriverUpdate {
  /** @minLength 2 */
  name?: string;
  /** @minLength 2 */
  shortName?: string;
  /** @minimum 0 */
  number?: number;
  /** @minLength 2 */
  team?: string;
  /** @minLength 2 */
  nationality?: string;
  rarity?: DriverUpdateRarity;
  /**
     * @minimum 1
     * @maximum 100
     */
  rating?: number;
  active?: boolean;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  spawnImageUrl?: string | null;
  /** @minimum 0 */
  spawnWeight?: number;
}